% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate.R
\name{byte_trunc}
\alias{byte_trunc}
\title{Truncate a character string by bytesize.}
\usage{
byte_trunc(string, size, ellipsis = "...")
}
\arguments{
\item{string}{A character vector.}

\item{size}{Maximum bytesize of string.}

\item{ellipsis}{Content of ellipsis that indicates content has been removed.}
}
\description{
Truncate a character string by bytesize.
}
\examples{
x <- "This string is moderately long"
byte_trunc(x, 20)

y <- "こんにちは"
byte_trunc(y, 10)
}
