% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bytetrunc2-package.R
\docType{package}
\name{bytetrunc2-package}
\alias{bytetrunc2}
\alias{bytetrunc2-package}
\title{bytetrunc2: Truncate text by bytesize}
\description{
Truncate text by a maximum bytesize in UTF-8, rather than by character count.
}
\author{
\strong{Maintainer}: Shun Sambongi \email{shunsambongi@gmail.com}

}
\keyword{internal}
